Progressing with the layout of the post
Added a few sections and prepared the motor_housing section
as I wouldn't want to write a too long of a post.
As a goal it would be nice to keep every entry to about 15 minutes
or read time and no more.
Additionally the motor housing itself was quite effortefull.